+++ packages/testing/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed an issue where `find()` and `findAll()` couldn't find components with fragment props.
+
 ## [1.0.3] - 2020-06-24
 
 ### Fixed